refactor: Moved functions and variables from dataManager to their respective places: LocalNotification, Common, and Constant.
